@@ -1,8 +1,6 @@
 module checkers-backend
 
-go 1.23.0
-
-toolchain go1.24.1
+go 1.24
 
 require (
 	// github.com/goccy/go-json v0.10.3
