@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.2
-// 	protoc        v5.27.2
+// 	protoc-gen-go v1.36.5
+// 	protoc        v6.30.1
 // source: base_payload.proto
 
 package game
@@ -11,6 +11,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -70,13 +71,10 @@ func (TeamColor) EnumDescriptor() ([]byte, []int) {
 }
 
 type BasePayload struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// GUI display text from server to client
 	Notice string `protobuf:"bytes,2,opt,name=notice,proto3" json:"notice,omitempty"`
-	// Types that are assignable to Inner:
+	// Types that are valid to be assigned to Inner:
 	//
 	//	*BasePayload_Welcome
 	//	*BasePayload_MovePayload
@@ -84,16 +82,16 @@ type BasePayload struct {
 	//	*BasePayload_ExitPayload
 	//	*BasePayload_CapturePayload
 	//	*BasePayload_WinlosePayload
-	Inner isBasePayload_Inner `protobuf_oneof:"inner"`
+	Inner         isBasePayload_Inner `protobuf_oneof:"inner"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *BasePayload) Reset() {
 	*x = BasePayload{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_base_payload_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_base_payload_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *BasePayload) String() string {
@@ -104,7 +102,7 @@ func (*BasePayload) ProtoMessage() {}
 
 func (x *BasePayload) ProtoReflect() protoreflect.Message {
 	mi := &file_base_payload_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -126,51 +124,63 @@ func (x *BasePayload) GetNotice() string {
 	return ""
 }
 
-func (m *BasePayload) GetInner() isBasePayload_Inner {
-	if m != nil {
-		return m.Inner
+func (x *BasePayload) GetInner() isBasePayload_Inner {
+	if x != nil {
+		return x.Inner
 	}
 	return nil
 }
 
 func (x *BasePayload) GetWelcome() *WelcomePayload {
-	if x, ok := x.GetInner().(*BasePayload_Welcome); ok {
-		return x.Welcome
+	if x != nil {
+		if x, ok := x.Inner.(*BasePayload_Welcome); ok {
+			return x.Welcome
+		}
 	}
 	return nil
 }
 
 func (x *BasePayload) GetMovePayload() *MovePayload {
-	if x, ok := x.GetInner().(*BasePayload_MovePayload); ok {
-		return x.MovePayload
+	if x != nil {
+		if x, ok := x.Inner.(*BasePayload_MovePayload); ok {
+			return x.MovePayload
+		}
 	}
 	return nil
 }
 
 func (x *BasePayload) GetStart() *StartPayload {
-	if x, ok := x.GetInner().(*BasePayload_Start); ok {
-		return x.Start
+	if x != nil {
+		if x, ok := x.Inner.(*BasePayload_Start); ok {
+			return x.Start
+		}
 	}
 	return nil
 }
 
 func (x *BasePayload) GetExitPayload() *ExitPayload {
-	if x, ok := x.GetInner().(*BasePayload_ExitPayload); ok {
-		return x.ExitPayload
+	if x != nil {
+		if x, ok := x.Inner.(*BasePayload_ExitPayload); ok {
+			return x.ExitPayload
+		}
 	}
 	return nil
 }
 
 func (x *BasePayload) GetCapturePayload() *CapturePayload {
-	if x, ok := x.GetInner().(*BasePayload_CapturePayload); ok {
-		return x.CapturePayload
+	if x != nil {
+		if x, ok := x.Inner.(*BasePayload_CapturePayload); ok {
+			return x.CapturePayload
+		}
 	}
 	return nil
 }
 
 func (x *BasePayload) GetWinlosePayload() *WinLosePayload {
-	if x, ok := x.GetInner().(*BasePayload_WinlosePayload); ok {
-		return x.WinlosePayload
+	if x != nil {
+		if x, ok := x.Inner.(*BasePayload_WinlosePayload); ok {
+			return x.WinlosePayload
+		}
 	}
 	return nil
 }
@@ -216,23 +226,20 @@ func (*BasePayload_CapturePayload) isBasePayload_Inner() {}
 func (*BasePayload_WinlosePayload) isBasePayload_Inner() {}
 
 type WelcomePayload struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// Color of this player pieces
 	MyTeam TeamColor `protobuf:"varint,1,opt,name=my_team,json=myTeam,proto3,enum=chk.payload.TeamColor" json:"my_team,omitempty"`
 	// in the form of X.Y.Z
 	ServerVersion string `protobuf:"bytes,2,opt,name=server_version,json=serverVersion,proto3" json:"server_version,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *WelcomePayload) Reset() {
 	*x = WelcomePayload{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_base_payload_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_base_payload_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *WelcomePayload) String() string {
@@ -243,7 +250,7 @@ func (*WelcomePayload) ProtoMessage() {}
 
 func (x *WelcomePayload) ProtoReflect() protoreflect.Message {
 	mi := &file_base_payload_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -273,23 +280,20 @@ func (x *WelcomePayload) GetServerVersion() string {
 }
 
 type StartPayload struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// Pieces for red player
 	PiecesRed []int32 `protobuf:"varint,1,rep,packed,name=pieces_red,json=piecesRed,proto3" json:"pieces_red,omitempty"`
 	// pieces for black player
-	PiecesBlack []int32 `protobuf:"varint,2,rep,packed,name=pieces_black,json=piecesBlack,proto3" json:"pieces_black,omitempty"`
+	PiecesBlack   []int32 `protobuf:"varint,2,rep,packed,name=pieces_black,json=piecesBlack,proto3" json:"pieces_black,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *StartPayload) Reset() {
 	*x = StartPayload{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_base_payload_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_base_payload_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StartPayload) String() string {
@@ -300,7 +304,7 @@ func (*StartPayload) ProtoMessage() {}
 
 func (x *StartPayload) ProtoReflect() protoreflect.Message {
 	mi := &file_base_payload_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -330,10 +334,7 @@ func (x *StartPayload) GetPiecesBlack() []int32 {
 }
 
 type MovePayload struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// from which player
 	FromTeam TeamColor `protobuf:"varint,1,opt,name=from_team,json=fromTeam,proto3,enum=chk.payload.TeamColor" json:"from_team,omitempty"`
 	// moving piece id
@@ -341,16 +342,16 @@ type MovePayload struct {
 	// older cell index for this pieceId
 	SourceCell int32 `protobuf:"varint,3,opt,name=source_cell,json=sourceCell,proto3" json:"source_cell,omitempty"`
 	// destination cell
-	Destination *MovePayload_Detination `protobuf:"bytes,4,opt,name=destination,proto3" json:"destination,omitempty"`
+	Destination   *MovePayload_Detination `protobuf:"bytes,4,opt,name=destination,proto3" json:"destination,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MovePayload) Reset() {
 	*x = MovePayload{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_base_payload_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_base_payload_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MovePayload) String() string {
@@ -361,7 +362,7 @@ func (*MovePayload) ProtoMessage() {}
 
 func (x *MovePayload) ProtoReflect() protoreflect.Message {
 	mi := &file_base_payload_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -406,21 +407,18 @@ func (x *MovePayload) GetDestination() *MovePayload_Detination {
 
 // whenever any of the players exits, or server terminates Match
 type ExitPayload struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// which player initiated exit (RED, BLACK, or unspecified)
-	FromTeam TeamColor `protobuf:"varint,1,opt,name=from_team,json=fromTeam,proto3,enum=chk.payload.TeamColor" json:"from_team,omitempty"`
+	FromTeam      TeamColor `protobuf:"varint,1,opt,name=from_team,json=fromTeam,proto3,enum=chk.payload.TeamColor" json:"from_team,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *ExitPayload) Reset() {
 	*x = ExitPayload{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_base_payload_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_base_payload_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ExitPayload) String() string {
@@ -431,7 +429,7 @@ func (*ExitPayload) ProtoMessage() {}
 
 func (x *ExitPayload) ProtoReflect() protoreflect.Message {
 	mi := &file_base_payload_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -455,25 +453,22 @@ func (x *ExitPayload) GetFromTeam() TeamColor {
 
 // when hunter player is capturing opponent's piece
 type CapturePayload struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// from which player
 	FromTeam TeamColor `protobuf:"varint,1,opt,name=from_team,json=fromTeam,proto3,enum=chk.payload.TeamColor" json:"from_team,omitempty"`
 	// the attacking player's piece
 	HunterPieceId int32                             `protobuf:"varint,2,opt,name=hunter_piece_id,json=hunterPieceId,proto3" json:"hunter_piece_id,omitempty"`
 	Details       *CapturePayload_TargetDetails     `protobuf:"bytes,4,opt,name=details,proto3" json:"details,omitempty"`
 	Destination   *CapturePayload_HunterDestination `protobuf:"bytes,5,opt,name=destination,proto3" json:"destination,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CapturePayload) Reset() {
 	*x = CapturePayload{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_base_payload_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_base_payload_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CapturePayload) String() string {
@@ -484,7 +479,7 @@ func (*CapturePayload) ProtoMessage() {}
 
 func (x *CapturePayload) ProtoReflect() protoreflect.Message {
 	mi := &file_base_payload_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -527,23 +522,20 @@ func (x *CapturePayload) GetDestination() *CapturePayload_HunterDestination {
 	return nil
 }
 
-// * When one of player wins. Also, this ends the  match
+// When one of player wins. Also, this marks the end of the match
 type WinLosePayload struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// which player won
-	Winner TeamColor `protobuf:"varint,1,opt,name=winner,proto3,enum=chk.payload.TeamColor" json:"winner,omitempty"`
+	Winner        TeamColor `protobuf:"varint,1,opt,name=winner,proto3,enum=chk.payload.TeamColor" json:"winner,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *WinLosePayload) Reset() {
 	*x = WinLosePayload{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_base_payload_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_base_payload_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *WinLosePayload) String() string {
@@ -554,7 +546,7 @@ func (*WinLosePayload) ProtoMessage() {}
 
 func (x *WinLosePayload) ProtoReflect() protoreflect.Message {
 	mi := &file_base_payload_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -578,22 +570,19 @@ func (x *WinLosePayload) GetWinner() TeamColor {
 
 // Where will this player land on
 type MovePayload_Detination struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	CellIndex     int32                  `protobuf:"varint,1,opt,name=cell_index,json=cellIndex,proto3" json:"cell_index,omitempty"`
+	X             float32                `protobuf:"fixed32,2,opt,name=x,proto3" json:"x,omitempty"`
+	Y             float32                `protobuf:"fixed32,3,opt,name=y,proto3" json:"y,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	CellIndex int32   `protobuf:"varint,1,opt,name=cell_index,json=cellIndex,proto3" json:"cell_index,omitempty"`
-	X         float32 `protobuf:"fixed32,2,opt,name=x,proto3" json:"x,omitempty"`
-	Y         float32 `protobuf:"fixed32,3,opt,name=y,proto3" json:"y,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MovePayload_Detination) Reset() {
 	*x = MovePayload_Detination{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_base_payload_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_base_payload_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MovePayload_Detination) String() string {
@@ -604,7 +593,7 @@ func (*MovePayload_Detination) ProtoMessage() {}
 
 func (x *MovePayload_Detination) ProtoReflect() protoreflect.Message {
 	mi := &file_base_payload_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -642,25 +631,22 @@ func (x *MovePayload_Detination) GetY() float32 {
 
 // extra details about this capture
 type CapturePayload_TargetDetails struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	// * ID of the piece that is to be captured
+	state protoimpl.MessageState `protogen:"open.v1"`
+	// ID of the piece that is to be captured
 	PreyPieceId int32 `protobuf:"varint,1,opt,name=prey_piece_id,json=preyPieceId,proto3" json:"prey_piece_id,omitempty"`
 	// the cell Index hosting this target piece
 	PreyCellIdx int32 `protobuf:"varint,2,opt,name=prey_cell_idx,json=preyCellIdx,proto3" json:"prey_cell_idx,omitempty"`
 	// source cell of hunter
 	HunterSrcCell int32 `protobuf:"varint,3,opt,name=hunter_src_cell,json=hunterSrcCell,proto3" json:"hunter_src_cell,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CapturePayload_TargetDetails) Reset() {
 	*x = CapturePayload_TargetDetails{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_base_payload_proto_msgTypes[8]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_base_payload_proto_msgTypes[8]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CapturePayload_TargetDetails) String() string {
@@ -671,7 +657,7 @@ func (*CapturePayload_TargetDetails) ProtoMessage() {}
 
 func (x *CapturePayload_TargetDetails) ProtoReflect() protoreflect.Message {
 	mi := &file_base_payload_proto_msgTypes[8]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -709,22 +695,19 @@ func (x *CapturePayload_TargetDetails) GetHunterSrcCell() int32 {
 
 // Destination of hunterPiece after capturing prey
 type CapturePayload_HunterDestination struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	CellIndex     int32                  `protobuf:"varint,1,opt,name=cell_index,json=cellIndex,proto3" json:"cell_index,omitempty"`
+	X             float32                `protobuf:"fixed32,2,opt,name=x,proto3" json:"x,omitempty"`
+	Y             float32                `protobuf:"fixed32,3,opt,name=y,proto3" json:"y,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	CellIndex int32   `protobuf:"varint,1,opt,name=cell_index,json=cellIndex,proto3" json:"cell_index,omitempty"`
-	X         float32 `protobuf:"fixed32,2,opt,name=x,proto3" json:"x,omitempty"`
-	Y         float32 `protobuf:"fixed32,3,opt,name=y,proto3" json:"y,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CapturePayload_HunterDestination) Reset() {
 	*x = CapturePayload_HunterDestination{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_base_payload_proto_msgTypes[9]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_base_payload_proto_msgTypes[9]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CapturePayload_HunterDestination) String() string {
@@ -735,7 +718,7 @@ func (*CapturePayload_HunterDestination) ProtoMessage() {}
 
 func (x *CapturePayload_HunterDestination) ProtoReflect() protoreflect.Message {
 	mi := &file_base_payload_proto_msgTypes[9]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -773,7 +756,7 @@ func (x *CapturePayload_HunterDestination) GetY() float32 {
 
 var File_base_payload_proto protoreflect.FileDescriptor
 
-var file_base_payload_proto_rawDesc = []byte{
+var file_base_payload_proto_rawDesc = string([]byte{
 	0x0a, 0x12, 0x62, 0x61, 0x73, 0x65, 0x5f, 0x70, 0x61, 0x79, 0x6c, 0x6f, 0x61, 0x64, 0x2e, 0x70,
 	0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0b, 0x63, 0x68, 0x6b, 0x2e, 0x70, 0x61, 0x79, 0x6c, 0x6f, 0x61,
 	0x64, 0x22, 0xa8, 0x03, 0x0a, 0x0b, 0x42, 0x61, 0x73, 0x65, 0x50, 0x61, 0x79, 0x6c, 0x6f, 0x61,
@@ -876,16 +859,16 @@ var file_base_payload_proto_rawDesc = []byte{
 	0x42, 0x4c, 0x41, 0x43, 0x4b, 0x10, 0x02, 0x42, 0x17, 0x5a, 0x15, 0x63, 0x68, 0x65, 0x63, 0x6b,
 	0x65, 0x72, 0x73, 0x2d, 0x62, 0x61, 0x63, 0x6b, 0x65, 0x6e, 0x64, 0x2f, 0x67, 0x61, 0x6d, 0x65,
 	0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+})
 
 var (
 	file_base_payload_proto_rawDescOnce sync.Once
-	file_base_payload_proto_rawDescData = file_base_payload_proto_rawDesc
+	file_base_payload_proto_rawDescData []byte
 )
 
 func file_base_payload_proto_rawDescGZIP() []byte {
 	file_base_payload_proto_rawDescOnce.Do(func() {
-		file_base_payload_proto_rawDescData = protoimpl.X.CompressGZIP(file_base_payload_proto_rawDescData)
+		file_base_payload_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_base_payload_proto_rawDesc), len(file_base_payload_proto_rawDesc)))
 	})
 	return file_base_payload_proto_rawDescData
 }
@@ -932,128 +915,6 @@ func file_base_payload_proto_init() {
 	if File_base_payload_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_base_payload_proto_msgTypes[0].Exporter = func(v any, i int) any {
-			switch v := v.(*BasePayload); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_base_payload_proto_msgTypes[1].Exporter = func(v any, i int) any {
-			switch v := v.(*WelcomePayload); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_base_payload_proto_msgTypes[2].Exporter = func(v any, i int) any {
-			switch v := v.(*StartPayload); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_base_payload_proto_msgTypes[3].Exporter = func(v any, i int) any {
-			switch v := v.(*MovePayload); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_base_payload_proto_msgTypes[4].Exporter = func(v any, i int) any {
-			switch v := v.(*ExitPayload); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_base_payload_proto_msgTypes[5].Exporter = func(v any, i int) any {
-			switch v := v.(*CapturePayload); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_base_payload_proto_msgTypes[6].Exporter = func(v any, i int) any {
-			switch v := v.(*WinLosePayload); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_base_payload_proto_msgTypes[7].Exporter = func(v any, i int) any {
-			switch v := v.(*MovePayload_Detination); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_base_payload_proto_msgTypes[8].Exporter = func(v any, i int) any {
-			switch v := v.(*CapturePayload_TargetDetails); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_base_payload_proto_msgTypes[9].Exporter = func(v any, i int) any {
-			switch v := v.(*CapturePayload_HunterDestination); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	file_base_payload_proto_msgTypes[0].OneofWrappers = []any{
 		(*BasePayload_Welcome)(nil),
 		(*BasePayload_MovePayload)(nil),
@@ -1066,7 +927,7 @@ func file_base_payload_proto_init() {
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_base_payload_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_base_payload_proto_rawDesc), len(file_base_payload_proto_rawDesc)),
 			NumEnums:      1,
 			NumMessages:   10,
 			NumExtensions: 0,
@@ -1078,7 +939,6 @@ func file_base_payload_proto_init() {
 		MessageInfos:      file_base_payload_proto_msgTypes,
 	}.Build()
 	File_base_payload_proto = out.File
-	file_base_payload_proto_rawDesc = nil
 	file_base_payload_proto_goTypes = nil
 	file_base_payload_proto_depIdxs = nil
 }
